Sanitize disconnect behavior
The disconnect funtion does not eat the connection object anyore, this
will allow to keep the connection object alive after disconnection.

This is useful since disconnection can happen by calling "close()" on
the connection or by a connection timeout. The timeout case keeps
the object alive so by symetry "close()" should as well.